Add comments on how a-scene needs to be loaded
See #3001 for the problem I found with some code that didn't run in Chrome. Thanks to @dmarcos and @ngokevin, who showed the need to wait until `a-scene` is loaded.

This section of the tutorial is the first one entering into details of how to write JavaScript code to manipulate a scene, so I think it needs these clarifications. Otherwise, you can lose a lot of time and energy until you realize your code should be in a listener for `loaded` of the `a-frame` element.